Add regulator job to all pipelines
This job is configured by default to run every 30 minutes. This job
always has a `get` configured for the `schedule` resource which has
`trigger: true` on it.

This job is additionally configured to do a get on the resource the
pipeline actually wants to track, for example pivnet-opsmgr. This is
what ties it into the rest of the pipeline, and how the rest of the
pipeline uses this as a gate.

This design follows the issue I created here:

https://github.com/concourse/concourse/issues/945