Create release pushes to different resource
I thought I could be saved from having to do this with [ci skip], but it
turns out that is not the case. The dependent get when doing a put
causes Concourse to ignore the [ci skip] rule and see a new version
anyway.

See related issue: https://github.com/concourse/git-resource/issues/18

This reverts commit 4de38807a9927a5ff63a8314d756c158a562ef89.