Do not use variable expansion on json strings
Using "${VARIABLE}" (wrapped in quotes) results in whitespace
preservation, which in the case of JSON is not necessary, and in the
case of JSON embedded in YAML, is harmful.

In this case the curly braces aren't useful, as we're not doing any bash
operations on the variables, just printing them as they are, so those
are removed as well.

Instead, use $VARIABLE (without the quotes).