Add bump-* jobs to CI pipeline
The idea is that when we run `publish-release` we take whatever rc
version is currently present and do two things; make a release with the
final version of that rc (e.g. 1.23.5-rc.1 -> 1.23.5), then bump the
version to the next patch version automatically (-> 1.23.6). Next time
pcf-pipelines passes `test` we can decide to just run `publish-rc` like
before, and go through that as many times as we like.

If at some point we decide there are more than just bug fixes in the
changes since the last release, we can run any of the bump-* jobs to
denote that prior to running `publish-rc`. This will take, for example,
1.25.5-rc.15 to 1.26.0 (bump-minor) or 2.0.0 (bump-major), and
then publish-rc will add the first rc to that version (1.26.0-rc.1 or
2.0.0-rc.1) We shouldn't really need to run bump-patch, but we can if we want.

This means that, by default, if you don't manually bump, it'll proceed
along bumping the patch version automatically after each final release
goes out

[#144896723]