Actually use individual db credentials
Originally it was just one set of credentials which were passed to
Terraform to create the database instance.

Then the service-specific credentials were added as params, but
discarded because AWS iaas-specific-config script didn't consume them;
in fact it just retrieved the original credentials used to instantiate
the instance and used those for each service, meaning the CVE was not
actually fixed.